made structuring fixes
