+++ src/providers/SmellsViewProvider.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
-import { getStatusIcon, getStatusMessage } from '../utils/fileStatus';
 import { buildPythonTree } from '../utils/TreeStructureBuilder';
 import { getAcronymByMessageId } from '../utils/smellsData';
 import { normalizePath } from '../utils/normalizePath';
@@ -186,5 +185,62 @@ export class SmellTreeItem extends vscode.TreeItem {
         arguments: [smell.path, firstLine - 1],
       };
     }
+  }
+}
+
+export function getStatusIcon(status: string): vscode.ThemeIcon {
+  switch (status) {
+    case 'queued':
+      return new vscode.ThemeIcon(
+        'sync~spin',
+        new vscode.ThemeColor('charts.yellow'),
+      );
+    case 'passed':
+      return new vscode.ThemeIcon('pass', new vscode.ThemeColor('charts.green'));
+    case 'no_issues':
+      return new vscode.ThemeIcon('pass', new vscode.ThemeColor('charts.blue'));
+    case 'failed':
+      return new vscode.ThemeIcon('error', new vscode.ThemeColor('charts.red'));
+    case 'outdated':
+      return new vscode.ThemeIcon('warning', new vscode.ThemeColor('charts.orange'));
+    case 'server_down':
+      return new vscode.ThemeIcon(
+        'server-process',
+        new vscode.ThemeColor('charts.red'),
+      );
+    case 'refactoring':
+      return new vscode.ThemeIcon('robot', new vscode.ThemeColor('charts.purple'));
+    case 'accept-refactoring':
+      return new vscode.ThemeIcon('warning', new vscode.ThemeColor('charts.yellow'));
+    default:
+      return new vscode.ThemeIcon('circle-outline');
+  }
+}
+
+/**
+ * Retrieves the status message corresponding to the smell analysis state.
+ * @param status - The analysis status.
+ * @returns A descriptive status message.
+ */
+export function getStatusMessage(status: string): string {
+  switch (status) {
+    case 'queued':
+      return 'Analyzing Smells';
+    case 'passed':
+      return 'Smells Successfully Detected';
+    case 'failed':
+      return 'Error Detecting Smells';
+    case 'no_issues':
+      return 'No Smells Found';
+    case 'outdated':
+      return 'File Outdated - Needs Reanalysis';
+    case 'server_down':
+      return 'Server Unavailable';
+    case 'refactoring':
+      return 'Refactoring Currently Ongoing';
+    case 'accept-refactoring':
+      return 'Successfully Refactored - Needs Reanalysis';
+    default:
+      return 'Smells Not Yet Detected';
   }
 }
