+++ test/setup.ts
@@ -1,3 +1,7 @@
+import mockEnvConfig from './mocks/env-config-mock';
+
 jest.mock('vscode');
 
-jest.mock('./../src/utils/envConfig');
+jest.mock('../src/utils/envConfig', () => ({
+  envConfig: mockEnvConfig,
+}));
